update tourio app to include bonus material. Update example implementation to reflect bonus. Update to UI. Minor CSS update to make the app look fresher.

+++ sessions/recap-project-6/example-implementation/pages/places/[id]/index.js
@@ -2,6 +2,7 @@ import Link from "next/link";
 import { useRouter } from "next/router.js";
 import useSWR from "swr";
 import styled from "styled-components";
+import Comments from "../../../components/Comments.js";
 import { StyledLink } from "../../../components/StyledLink.js";
 import { StyledButton } from "../../../components/StyledButton.js";
 import { StyledImage } from "../../../components/StyledImage.js";
@@ -13,8 +14,9 @@ const ImageContainer = styled.div`
 
 const ButtonContainer = styled.section`
   display: flex;
+  flex-direction: column;
   justify-content: space-between;
-  gap: 0.2rem;
+  gap: 0.5rem;
 
   & > * {
     flex-grow: 1;
@@ -24,8 +26,9 @@ const ButtonContainer = styled.section`
 
 const StyledLocationLink = styled(StyledLink)`
   text-align: center;
-  background-color: white;
-  border: 3px solid lightsalmon;
+  background-color: lightgray;
+  color: black;
+  border: none;
 `;
 
 export default function DetailsPage() {
@@ -33,21 +36,26 @@ export default function DetailsPage() {
   const { isReady } = router;
   const { id } = router.query;
 
-  const { data: place, isLoading, error } = useSWR(`/api/places/${id}`);
-
+  const { data, mutate, isLoading, error } = useSWR(`/api/places/${id}`);
+  console.log("data", data);
   if (!isReady || isLoading || error) return <h2>Loading...</h2>;
 
   async function deletePlace() {
-    await fetch(`/api/places/${id}`, {
-      method: "DELETE",
-    });
-    router.push("/");
+    if (confirm("are you sure you want to delete this place?")) {
+      await fetch(`/api/places/${id}`, {
+        method: "DELETE",
+      });
+      router.push("/");
+      return;
+    }
   }
+
+  const { place, comments } = data;
 
   return (
     <>
       <Link href={"/"} passHref legacyBehavior>
-        <StyledLink $justifySelf="start">back</StyledLink>
+        <StyledLink justifySelf="start">back</StyledLink>
       </Link>
       <ImageContainer>
         <StyledImage
@@ -71,10 +79,11 @@ export default function DetailsPage() {
         <Link href={`/places/${id}/edit`} passHref legacyBehavior>
           <StyledLink>Edit</StyledLink>
         </Link>
-        <StyledButton onClick={deletePlace} type="button" $variant="delete">
+        <StyledButton onClick={deletePlace} type="button" variant="delete">
           Delete
         </StyledButton>
       </ButtonContainer>
+      <Comments locationName={place.name} mutate={mutate} comments={comments} />
     </>
   );
 }
